+++ dvc.lock
@@ -3,25 +3,25 @@ stages:
   download_and_extract_osm:
     cmd: wget -c --limit-rate=10M -P data/
       "https://download.geofabrik.de/russia/central-fed-district-latest.osm.pbf"
-      && osmium extract --bbox "37.37411,55.63187,37.60483,55.74006"
+      && osmium extract --bbox "37.44186,55.52806,37.57599,55.58824"
       data/central-fed-district-latest.osm.pbf --output data/map.osm.pbf
     params:
       params.yaml:
         prepare_satellite_data.map_box:
-        - 37.37411
-        - 55.63187
-        - 37.60483
-        - 55.74006
+        - 37.44186
+        - 55.52806
+        - 37.57599
+        - 55.58824
         prepare_satellite_data.osm: data/map.osm.pbf
     outs:
     - path: data/central-fed-district-latest.osm.pbf
       hash: md5
-      md5: ad9b04a8e07e904fa232d1f48ebaaee2
-      size: 858465154
+      md5: 1f96352eda6d631d537e232c5101f168
+      size: 858879847
     - path: data/map.osm.pbf
       hash: md5
-      md5: 67e4196db5425da2b05fc8c446dbb28b
-      size: 9654251
+      md5: f0bf6c5e45579a19d6915b7c480403a5
+      size: 1995562
   prepare_satellite_data:
     cmd: python scripts/prepare_satellite_data.py
     params:
@@ -39,10 +39,10 @@ stages:
           tile_size: 256
           dataset_dir: data/dataset
           map_box:
-          - 36.9
-          - 55.38
-          - 38.0
-          - 56.0
+          - 37.44186
+          - 55.52806
+          - 37.57599
+          - 55.58824
           osm: data/map.osm.pbf
           sources:
             google: http://mt1.google.com/vt/lyrs=s&x={x}&y={y}&z={z}
@@ -53,9 +53,9 @@ stages:
     outs:
     - path: data/satellite
       hash: md5
-      md5: daaf3ec14030c7ccb5e61d6791e1b313.dir
-      size: 9841615899
-      nfiles: 8006
+      md5: 3ee435178801824b59e888db1be52ce8.dir
+      size: 80278954
+      nfiles: 60
   prepare_dataset:
     cmd: python scripts/prepare_dataset.py
     params:
@@ -73,10 +73,10 @@ stages:
           tile_size: 256
           dataset_dir: data/dataset
           map_box:
-          - 36.9
-          - 55.38
-          - 38.0
-          - 56.0
+          - 37.44186
+          - 55.52806
+          - 37.57599
+          - 55.58824
           osm: data/map.osm.pbf
           sources:
             google: http://mt1.google.com/vt/lyrs=s&x={x}&y={y}&z={z}
@@ -87,32 +87,32 @@ stages:
     outs:
     - path: data/dataset
       hash: md5
-      md5: f2b584c8339a13f01aebb6eb36607627.dir
-      size: 147074803200
-      nfiles: 1600
+      md5: 9e893d94056640dc319077d620c86f29.dir
+      size: 1103061024
+      nfiles: 12
   training:
     cmd: python src/train.py
     deps:
     - path: data/dataset
       hash: md5
-      md5: f2b584c8339a13f01aebb6eb36607627.dir
-      size: 147074803200
-      nfiles: 1600
+      md5: 9e893d94056640dc319077d620c86f29.dir
+      size: 1103061024
+      nfiles: 12
     params:
       params.yaml:
         training:
           random_seed: 42
           checkpoint_dir: data/checkpoint
           classes: ${prepare_satellite_data.classes}
-          num_epoch: 25
+          num_epoch: 1
           loader:
             dataset_dir: ${prepare_dataset.dataset_dir}
             train_ratio: 0.7
             val_ratio: 0.15
             test_ratio: 0.15
-            batch_size: 32
+            batch_size: 1
             num_workers: 4
-            image_size: 512
+            image_size: 256
           model:
             loss: mIoU
             metric: mcc
